--- a/tests/utils.py
+++ b/tests/utils.py
@@ -121,7 +121,8 @@ simple_attrs = st.one_of(bare_attrs, int_attrs, str_attrs, float_attrs,
                          dict_attrs)
 
 # Python functions support up to 255 arguments.
-simple_classes = st.lists(simple_attrs, max_size=255).map(_create_hyp_class)
+simple_classes = (st.lists(simple_attrs, average_size=9, max_size=50)
+                  .map(_create_hyp_class))
 
 # Ok, so st.recursive works by taking a base strategy (in this case,
 # simple_classes) and a special function. This function receives a strategy,