@Tinche 2016-08-17
If we're up for this, converters can be given the same treatment.