@hynek 2016-08-18
I’m not gonna lie, this looks scary. 😱  But 50% faster is kind of sweet. 🙈

How would a non-public `_validate` that doesn’t do the safety stuff work out performance-wise?