@Tinche 2016-08-18
I can check later today.

I really like the fact that at class definition time we actually know which attributes have validators, and we can generate an `__init__` calling exactly those validators. It somehow feels it'd be a waste to not use this information, and it plays into our spiel of attrs generating classes better than or equal than you yourself can write.

(I'll fix the coverage too.)