@hynek 2016-08-18
Yeah you’re right.  I’m just being lazy about rebasing immutable.