@Tinche 2016-08-20
Dammit, only now I see apparently limiting the number of generated attributes breaks tests!?!

Will take a closer look when I have some time.