@hynek 2016-08-20
this needs a rebase :-P