@hynek 2016-08-20
also maybe the the average for the number of attributes to a low single digit please. :)  that should help with the slownesses…