@hynek 2016-08-22
We’re all going to hell but 16.1 is gonna be majestic.